Move TupleTypeDecl to a separate file
